+++ docs/changelog.rst
@@ -26,6 +26,7 @@ CHANGELOG
 - Add OpenStreetMap parser for infrastructures
 - Add OpenStreetMap parser for touristic contents
 - Add OpenStreetMap parser for Outdoor Sites
+- Change provider Charfield for a foreign key
 - All parsers for point topology objects now generate the topology
 - Add OpenStreetMap parser for Cities
 - Add Apidae parser for Infrastructure